releaseMergeFFOnly is not used when merging release to develop.
Typically the develop branch moves on while the release is being
prepared. Therefore it does not make sense to require the use of
fast-forward when merging to the develop branch.

On the master branch, however, it is useful to be able to require the
use of fast-forward to make sure that the final commit on the release
branch ends up in the master branch unmodified.