#230 After conflict solved should not fail on tagging or POM commit
After the user has solved the conflicts that appeared during the first
hotfix-finish, it should be possible to restart the hotfix-finish
without failing. The first fix is that if we are tagging (skipTag=false)
we should allow the user to avoid failing. Since the tagging happened on
master before possible merge fail on develop, we would fail when tagging
the second time around. The second failure fix is to make sure that the
development version is not incremented indiscriminately.
There is one last problem, which is not solved yet: the fact that if we
have a conflict when merging on develop, we should backtrack the commit
were we set the version to hotfix branch...