udpate versions (pom and property) in the same run
If versions:set and versions:set-property are launched in separate
maven commands, versions:set-property can fail as poms are in an
inconsistent state.

Attached test case shows a practical use-case where a property is used
in dependencyManagement declaration in a multi-module project.

If project is not installed, versions:set-property 1.0.1-SNAPSHOT ->
1.0.1 fails as maven tries to resolve dependency 1.0.1-SNAPSHOT in a
project already versionned in 1.0.1.